fruit-inference/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
ultralytics>=8.2.0
boto3>=1.34.0
pillow>=10.3.0
python-dotenv>=1.0.1
pydantic>=2.7.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
